@@ -14,42 +14,43 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.5",
-    "@angular/cdk": "^5.2.1",
-    "@angular/common": "^5.2.5",
-    "@angular/compiler": "^5.2.5",
-    "@angular/core": "^5.2.5",
-    "@angular/forms": "^5.2.5",
-    "@angular/http": "^5.2.5",
-    "@angular/material": "^5.2.1",
-    "@angular/platform-browser": "^5.2.5",
-    "@angular/platform-browser-dynamic": "^5.2.5",
-    "@angular/router": "^5.2.5",
-    "core-js": "^2.5.3",
+    "@angular/animations": "^6.0.0",
+    "@angular/cdk": "^6.0.1",
+    "@angular/common": "^6.0.0",
+    "@angular/compiler": "^6.0.0",
+    "@angular/core": "^6.0.0",
+    "@angular/forms": "^6.0.0",
+    "@angular/http": "^6.0.0",
+    "@angular/material": "^6.0.1",
+    "@angular/platform-browser": "^6.0.0",
+    "@angular/platform-browser-dynamic": "^6.0.0",
+    "@angular/router": "^6.0.0",
+    "core-js": "^2.5.6",
     "hammerjs": "^2.0.8",
-    "rxjs": "^5.5.6",
-    "zone.js": "^0.8.20"
+    "rxjs": "^6.1.0",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.7.0",
-    "@angular/compiler-cli": "^5.2.5",
+    "@angular/cli": "^6.0.0",
+    "@angular/compiler-cli": "^6.0.0",
     "@types/hammerjs": "^2.0.35",
-    "@types/jasmine": "2.8.2",
-    "@types/node": "~9.4.6",
-    "codelyzer": "~4.1.0",
-    "jasmine-core": "~2.99.1",
+    "@types/jasmine": "2.8.7",
+    "@types/node": "~10.0.4",
+    "codelyzer": "~4.3.0",
+    "jasmine-core": "~3.1.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
+    "karma": "~2.0.2",
     "karma-chrome-launcher": "~2.2.0",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.4.1",
-    "karma-jasmine": "~1.1.1",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-coverage-istanbul-reporter": "^1.4.2",
+    "karma-jasmine": "~1.1.2",
+    "karma-jasmine-html-reporter": "^1.0.0",
     "karma-phantomjs-launcher": "^1.0.4",
     "phantomjs-prebuilt": "^2.1.16",
-    "protractor": "~5.3.0",
-    "ts-node": "~4.1.0",
-    "tslint": "~5.9.1",
-    "typescript": "~2.7.2"
+    "protractor": "~5.3.1",
+    "ts-node": "~6.0.3",
+    "tslint": "~5.10.0",
+    "typescript": "~2.7.0",
+    "@angular-devkit/build-angular": "~0.6.0"
   }
 }
